+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## 12.0.0 (unreleased)
 
+Features:
+
+- `Env.enum` allows passing an `Enum` as `default`.
+
+Other changes:
+
 - Support marshmallow 4.
 - Support Python 3.9-3.13.
 - _Backwards-incompatible_: Make `Field` classes private. Users should